requirements.txt
supabase
langchain-huggingface
langchain-chroma
python-dotenv
requests
Pillow
sentence-transformers.venev